Removed no longer needed UInt8VertexBufferEnabled flag from AppState, in the renderer of the OXR version - turns out, support for 8-bit RGBA values is mandatory in Vulkan for vertex input data. Modified alias & viewmodel vertex shaders to use always uvec4 for vertex position input. Updated Kotlin version to 1.6.21 .

+++ SlipNFrag-XR/app/src/main/shaders/alias.vert
@@ -16,7 +16,7 @@ layout(push_constant) uniform Transforms
 	layout(offset = 0) mat4 aliasTransform;
 };
 
-layout(location = 0) in vec4 vertexPosition;
+layout(location = 0) in uvec4 vertexPosition;
 layout(location = 1) in vec2 vertexTexCoords;
 layout(location = 2) in float vertexLight;
 layout(location = 0) out mediump vec3 fragmentData;
